# openapi.yaml
openapi: 3.0.3
info:
  title: Kopahi API
  version: 1.0.0
  description: |
    REST API for the Kopahi B2B2C agri-marketplace.
    All authenticated endpoints accept a JWT in `Authorization: Bearer <token>`.
servers:
  - url: http://localhost:5000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/orders:
    post:
      tags: [Orders]
      security: [{ bearerAuth: [] }]
      summary: Place order. Server recomputes prices, applies coupon, decrements stock.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      product: { type: string }
                      quantity: { type: integer }
                shippingAddress:
                  type: object
                paymentMethod: { type: string, enum: [COD, Online] }
                couponCode: { type: string }
      responses:
        "201": { description: Created }
    get:
      tags: [Orders]
      security: [{ bearerAuth: [] }]
      summary: Admin — list all orders
      responses:
        "200": { description: OK }
  /api/orders/mine:
    get:
      tags: [Orders]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
  /api/orders/{id}:
    get:
      tags: [Orders]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
  /api/orders/{id}/status:
    put:
      tags: [Orders]
      security: [{ bearerAuth: [] }]
      summary: Admin — update order status
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
  /api/orders/{id}/cancel:
    put:
      tags: [Orders]
      security: [{ bearerAuth: [] }]
      summary: Cancel order (owner or admin). Restores stock.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
  /api/orders/coupon/validate:
    post:
      tags: [Orders]
      summary: Validate a coupon code against a subtotal
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                code: { type: string }
                subtotal: { type: number }
      responses:
        "200": { description: OK }
        "400": { description: Invalid }
...
  /api/coupons:
    get:
      tags: [Coupons]
      security: [{ bearerAuth: [] }]
      summary: Admin — list coupons
      responses:
        "200": { description: OK }
    post:
      tags: [Coupons]
      security: [{ bearerAuth: [] }]
      summary: Admin — create coupon
      responses:
        "201": { description: Created }
  /api/blog:
    get:
      tags: [Blog]
      responses:
        "200": { description: OK }
    post:
      tags: [Blog]
      security: [{ bearerAuth: [] }]
      summary: Admin — create blog post
      responses:
        "201": { description: Created }
  /api/blog/{slug}:
    get:
      tags: [Blog]
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
...
  /api/admin/dashboard:
    get:
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
  /api/admin/audit:
    get:
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      summary: Recent audit log entries
      responses:
        "200": { description: OK }
...
  /api/payment/razorpay/order:
    post:
      tags: [Payment]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
        "503": { description: Payments not configured }
...